channels , ExchangeRates, InventorySources, TaxCategory, TaxRates Done

+++ packages/Webkul/Admin/src/DataGrids/Settings/ChannelDataGrid.php
@@ -76,22 +76,26 @@ class ChannelDataGrid extends DataGrid
 
     public function prepareActions()
     {
-        $this->addAction([
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.settings.channels.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => function ($row) {
-                return route('admin.settings.channels.edit', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.channels.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.settings.channels.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => function ($row) {
+                    return route('admin.settings.channels.edit', $row->id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.settings.channels.index.datagrid.delete'),
-            'method' => 'DELETE',
-            'url'    => function ($row) {
-                return route('admin.settings.channels.delete', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.channels.delete')) {
+            $this->addAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.settings.channels.index.datagrid.delete'),
+                'method' => 'DELETE',
+                'url'    => function ($row) {
+                    return route('admin.settings.channels.delete', $row->id);
+                },
+            ]);
+        }
     }
 }
