+++ packages/Webkul/Admin/src/DataGrids/Settings/ExchangeRatesDataGrid.php
@@ -68,22 +68,26 @@ class ExchangeRatesDataGrid extends DataGrid
 
     public function prepareActions()
     {
-        $this->addAction([
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.settings.exchange-rates.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => function ($row) {
-                return route('admin.settings.exchange_rates.edit', $row->currency_exchange_id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.exchange_rates.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.settings.exchange-rates.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => function ($row) {
+                    return route('admin.settings.exchange_rates.edit', $row->currency_exchange_id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.settings.exchange-rates.index.datagrid.delete'),
-            'method' => 'DELETE',
-            'url'    => function ($row) {
-                return route('admin.settings.exchange_rates.delete', $row->currency_exchange_id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.exchange_rates.delete')) {
+            $this->addAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.settings.exchange-rates.index.datagrid.delete'),
+                'method' => 'DELETE',
+                'url'    => function ($row) {
+                    return route('admin.settings.exchange_rates.delete', $row->currency_exchange_id);
+                },
+            ]);
+        }
     }
 }
