+++ packages/Webkul/Admin/src/DataGrids/Settings/InventorySourcesDataGrid.php
@@ -93,22 +93,26 @@ class InventorySourcesDataGrid extends DataGrid
      */
     public function prepareActions()
     {
-        $this->addAction([
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.settings.inventory-sources.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => function ($row) {
-                return route('admin.settings.inventory_sources.edit', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.inventory_sources.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.settings.inventory-sources.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => function ($row) {
+                    return route('admin.settings.inventory_sources.edit', $row->id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.settings.inventory-sources.index.datagrid.delete'),
-            'method' => 'DELETE',
-            'url'    => function ($row) {
-                return route('admin.settings.inventory_sources.delete', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.inventory_sources.delete')) {
+            $this->addAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.settings.inventory-sources.index.datagrid.delete'),
+                'method' => 'DELETE',
+                'url'    => function ($row) {
+                    return route('admin.settings.inventory_sources.delete', $row->id);
+                },
+            ]);
+        }
     }
 }
