+++ packages/Webkul/Admin/src/Resources/views/tax/categories/index.blade.php
@@ -13,12 +13,14 @@
             <div class="flex gap-x-[10px] items-center">
                 <div class="flex gap-x-[10px] items-center">
                     {{-- Create Tax Category Button --}}
-                    <button
-                        type="button"
-                        class="primary-button"
-                    >
-                        @lang('admin::app.settings.taxes.categories.index.create.title')
-                    </button>
+                    @if (bouncer()->hasPermission('settings.taxes.tax-categories.create'))
+                        <button
+                            type="button"
+                            class="primary-button"
+                        >
+                            @lang('admin::app.settings.taxes.categories.index.create.title')
+                        </button>
+                    @endif
                 </div>
             </div>
         </div>
@@ -37,7 +39,7 @@
                 <div class="flex gap-x-[10px] items-center">
                     <div class="flex gap-x-[10px] items-center">
                         <!-- Create Tax Category Button -->
-                        @if (bouncer()->hasPermission('settings.taxes.categories.create'))
+                        @if (bouncer()->hasPermission('settings.taxes.tax-categories.create'))
                             <button
                                 type="button"
                                 class="primary-button"
@@ -55,9 +57,13 @@
                 :src="route('admin.settings.taxes.categories.index')"
                 ref="datagrid"
             >
+                @php
+                    $hasPermission = bouncer()->hasPermission('settings.taxes.tax-categories.edit') || bouncer()->hasPermission('settings.taxes.tax-categories.delete');
+                @endphp
+
                 <!-- DataGrid Header -->
                 <template #header="{ columns, records, sortPage, applied}">
-                    <div class="row grid grid-cols-4 grid-rows-1 gap-[10px] items-center px-[16px] py-[10px] border-b-[1px] border-gray-300 text-gray-600 bg-gray-50 font-semibold">
+                    <div class="row grid grid-cols-{{ $hasPermission ? '4' : '3' }} grid-rows-1 gap-[10px] items-center px-[16px] py-[10px] border-b-[1px] border-gray-300 text-gray-600 bg-gray-50 font-semibold">
                         <div
                             class="flex gap-[10px] cursor-pointer"
                             v-for="(columnGroup, index) in ['id', 'code', 'name']"
@@ -88,9 +94,11 @@
                         </div>
 
                         <!-- Actions -->
-                        <p class="flex gap-[10px] justify-end">
-                            @lang('admin::app.components.datagrid.table.actions')
-                        </p>
+                        @if ($hasPermission)
+                            <p class="flex gap-[10px] justify-end">
+                                @lang('admin::app.components.datagrid.table.actions')
+                            </p>
+                        @endif
                     </div>
                 </template>
 
@@ -99,7 +107,7 @@
                     <div
                         v-for="record in records"
                         class="row grid gap-[10px] items-center px-[16px] py-[16px] border-b-[1px] border-gray-300 text-gray-600 transition-all hover:bg-gray-50"
-                        style="grid-template-columns: repeat(4, 1fr);"
+                        :style="'grid-template-columns: repeat(' + (record.actions.length ? 4 : 3) + ', 1fr);'"
                     >
                         <!-- Id -->
                         <p v-text="record.id"></p>
@@ -113,23 +121,27 @@
 
                         <!-- Actions -->
                         <div class="flex justify-end">
-                            <a @click="id=1; editModal(record.id)">
-                                <span
-                                    :class="record.actions['0'].icon"
-                                    class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-100 max-sm:place-self-center"
-                                    :title="record.actions['0'].title"
-                                >
-                                </span>
-                            </a>
+                            @if (bouncer()->hasPermission('settings.taxes.tax-categories.edit'))
+                                <a @click="id=1; editModal(record.id)">
+                                    <span
+                                        :class="record.actions['0'].icon"
+                                        class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-100 max-sm:place-self-center"
+                                        :title="record.actions['0'].title"
+                                    >
+                                    </span>
+                                </a>
+                            @endif
 
-                            <a @click="deleteModal(record.actions['1']?.url)">
-                                <span
-                                    :class="record.actions['1'].icon"
-                                    class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-100 max-sm:place-self-center"
-                                    :title="record.actions['1'].title"
-                                >
-                                </span>
-                            </a>
+                            @if (bouncer()->hasPermission('settings.taxes.tax-categories.delete'))
+                                <a @click="deleteModal(record.actions['1']?.url)">
+                                    <span
+                                        :class="record.actions['1'].icon"
+                                        class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-100 max-sm:place-self-center"
+                                        :title="record.actions['1'].title"
+                                    >
+                                    </span>
+                                </a>
+                            @endif
                         </div>
                     </div>
                 </template>
